Update README.md
Although, I'm not sure of the importance lol. I felt it was necessary to let the person know that since they are porting their Roblox project, which has pretty loose naming requirements into folders. They should know that there are a few extra restrictions on how they should be naming any specific instance within their project because it all needs to be ported into folders, to then be used professionally with Rojo.